- add inventory.model.js - add inventory parsing functionality
