+++ server/src/api/controllers/product.controller.js
@@ -2,11 +2,13 @@ const httpStatus = require('http-status');
 const {omit, map} = require('lodash');
 const File = require('../models/file.model');
 const Product = require('../models/product.model');
+const Inventory = require('../models/inventory.model');
 const path = require('path');
 const fs = require('fs');
 const APIError = require('../utils/APIError');
 const parser = require('xml2json');
 const csv = require('csvtojson')
+const {MapToProductObject} = require('../utils/ProductMapper')
 /**
  * Upload Product xml or inventory csv files
  * @public
@@ -29,7 +31,6 @@ exports.upload = async (req, res, next) => {
             }
         });
 
-        // TODO Save xml or csv file and do the logic
         if (mimeType === 'application/xml' || mimeType === 'text/xml') {
 
             fs.readFile(filePath, function (err, data) {
@@ -53,10 +54,19 @@ exports.upload = async (req, res, next) => {
                 delimiter: [";"]
             }).fromFile(filePath);
 
-            /*var mappedCsv = map(jsonArray, (csvItem) => {
-                csvItem.amount = parseFloat(csvItem.amount.replace(",", "."));
-                Product.updateOne({handle: csvItem.handle}, {$set: {amount:  csvItem.amount}})
-            });*/
+            jsonArray.forEach(inventory => {
+                inventory.amount = inventory.amount.replace(',','.');
+                Inventory.updateOne(
+                    {
+                        handle: inventory.handle,
+                        location: inventory.location
+                    },
+                    inventory,
+                    {
+                        upsert: true,
+                        setDefaultsOnInsert: true
+                    }).exec();
+            });
         }
 
         await File.create(fileObject, (err) => {
@@ -133,29 +143,4 @@ exports.uploadedFile = async (req, res, next) => {
     }
 };
 
-function MapToProductObject(products) {
-    return map(products, (product) => {
-            return {
-                id: product['id']['$t'],
-                title: product['title']['$t'],
-                bodyHtml: product['body-html']['$t'],
-                vendor: product['vendor']['$t'],
-                productType: product['product-type']['$t'],
-                createdAt: product['created-at']['$t'],
-                handle: product['handle']['$t'],
-                publishedScope: product['published-scope']['$t'],
-                tags: product['tags']['$t'],
-                image: {
-                    id: product['image']['id']['$t'],
-                    productId: product['image']['product-id']['$t'],
-                    createdAt: product['image']['created-at']['$t'],
-                    updatedAt: product['image']['updated-at']['$t'],
-                    alt: product.image?.alt?.nil,
-                    width: product['image']['width']['$t'],
-                    height: product['image']['height']['$t'],
-                    src: product['image']['src']['$t'],
-                }
-            };
-        }
-    );
-}
+
